+++ internal/app/auth/delivery/implementation.go
@@ -1,27 +1,28 @@
-package api
+package delivery
 
 import (
 	"context"
 
 	"github.com/go-park-mail-ru/2024_2_GOATS/internal/app/models"
+	"github.com/go-park-mail-ru/2024_2_GOATS/internal/app/models/auth"
 	authModels "github.com/go-park-mail-ru/2024_2_GOATS/internal/app/models/auth"
 )
 
-type ServiceInterface interface {
+type AuthServiceInterface interface {
 	Login(ctx context.Context, loginData *authModels.LoginData) (*authModels.AuthResponse, *models.ErrorResponse)
 	Register(ctx context.Context, registerData *authModels.RegisterData) (*authModels.AuthResponse, *models.ErrorResponse)
-	GetCollection(ctx context.Context) (*models.CollectionsResponse, *models.ErrorResponse)
 	Session(ctx context.Context, cookie string) (*authModels.SessionResponse, *models.ErrorResponse)
+	Logout(ctx context.Context, cookie string) (*auth.AuthResponse, *models.ErrorResponse)
 }
 
 type Implementation struct {
 	Ctx     context.Context
-	service ServiceInterface
+	authService AuthServiceInterface
 }
 
-func NewImplementation(ctx context.Context, srv ServiceInterface) *Implementation {
+func NewImplementation(ctx context.Context, srv AuthServiceInterface) *Implementation {
 	return &Implementation{
 		Ctx:     ctx,
-		service: srv,
+		authService: srv,
 	}
 }
